Refactor epoch processing test helpers
