+++ tests/core/pyspec/eth2spec/test/lightclient_patch/epoch_processing/test_process_final_updates.py
@@ -4,7 +4,7 @@ from eth2spec.test.context import (
     spec_state_test,
 )
 from eth2spec.test.helpers.state import transition_to
-from eth2spec.test.phase0.epoch_processing.run_epoch_process_base import (
+from eth2spec.test.helpers.epoch_processing import (
     run_epoch_processing_with,
 )
