+++ tests/core/pyspec/eth2spec/test/phase0/epoch_processing/test_process_justification_and_finalization.py
@@ -1,5 +1,5 @@
 from eth2spec.test.context import spec_state_test, with_all_phases
-from eth2spec.test.phase0.epoch_processing.run_epoch_process_base import (
+from eth2spec.test.helpers.epoch_processing import (
     run_epoch_processing_with
 )
 from eth2spec.test.helpers.state import transition_to
